requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.0
mysqlclient==2.2.4
platformdirs==4.3.2
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.4
